Fixed Paths so images and CSS would work
Good job! When you moved over your code though, the url changed a bit for your site so I had to fix the paths to your images and CSS.
The old way you had it was like:
/images/whatever.jpg -> this is short for http://whateverdomain.com/images/whatever.jpg
(notice how "images" comes RIGHT AFTER the domain name in the URL?)
The / tells it to move up to the top level folder after the domain name (whateverdomain.com).
With the way urls were setup on c9 this worked...but on github the URL is setup a bit different
our project name comes after the domain name, then starts our files...
so it's like: http://my-username.github.io/my-project/images/whatever.jpg
There are a few ways we can handle this...
Easiest is to use a RELATIVE path...meaning we check for the images folder as it relates to our index.html file
which is easy :-) we just take off the /
images/whatever.jpg (without /) - this says check the same directory as index.html for a folder named images then look inside it for whatever/jpg.